Refactor code to separate out resource handling and azure repo helpers, node js webapp smart handle
