+++ azure-devops/azext_devops/dev/pipelines/pipeline_create_helpers/__init__.py
@@ -1,0 +1,8 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+import pkg_resources
+
+pkg_resources.declare_namespace(__name__)
